requirements.txt
flask
flask-sqlalchemy
flask-smorest
flask-migrate
psycopg2
gunicorn
python-dotenv
requests

